添付ファイル保存モード追加

+++ src/java/db/AttachmentDb.java
@@ -1,6 +1,6 @@
 package db;
 
-import constant.AttachmentBaseDir;
+import constant.AttachmentPropaties;
 import entity.Attachment;
 import entity.AttachmentKey;
 import java.io.BufferedOutputStream;
@@ -43,44 +43,45 @@ public class AttachmentDb extends TryCatchDb<Attachment> {
         if (maxSeq != null) {
             nextSeq = maxSeq + 1;
         }
-        // 添付ファイルのデータをDBに登録する場合。
-//        try (InputStream in = part.getValueAs(InputStream.class)) {
-//            byte[] data = IOUtils.toByteArray(in);
-//            in.read(data);
-//            String filename = new String(part.getFormDataContentDisposition().getFileName().getBytes("iso-8859-1"), "utf-8");
-//            Attachment attachment = new Attachment(new AttachmentKey(id, nextSeq), filename, data);
-//            add(attachment);
-//            
-//            System.out.println("filename = " + filename + ", size = " + data.length);
-//        } catch (IOException ex) {
-//            Logger.getLogger(AttachmentDb.class.getName()).log(Level.SEVERE, null, ex);
-//        }
-        
-        // サーバのフォルダに保存する場合。
-        // ファイルを配置するフォルダを作成する。　ベースフォルダがない場合はベースフォルダごと作成する。
-        Path putDir = Paths.get(AttachmentBaseDir.BASE_DIR + id + "/" + nextSeq);
-        if (Files.notExists(putDir)) {
-            try {
-                Files.createDirectories(putDir);
+        // 添付ファイルモードによって処理を切り替える
+        if (AttachmentPropaties.ATTACHMENT_MODE == AttachmentPropaties.AttachmentMode.DB_MODE) {
+            // 添付ファイルのデータをDBに登録する場合。
+            try (InputStream in = part.getValueAs(InputStream.class)) {
+                byte[] data = IOUtils.toByteArray(in);
+                in.read(data);
+                String filename = new String(part.getFormDataContentDisposition().getFileName().getBytes("iso-8859-1"), "utf-8");
+                Attachment attachment = new Attachment(new AttachmentKey(id, nextSeq), filename, data);
+                add(attachment);
             } catch (IOException ex) {
                 Logger.getLogger(AttachmentDb.class.getName()).log(Level.SEVERE, null, ex);
             }
-        }
-        // ファイルをフォルダに配置する。
-        try (InputStream in = part.getValueAs(InputStream.class)) {
-            // DBにデータを登録する。（添付ファイルはDBには入れずにサーバに保存する。）
-            String filename = new String(part.getFormDataContentDisposition().getFileName().getBytes("iso-8859-1"), "utf-8");
-            Attachment attachment = new Attachment(new AttachmentKey(id, nextSeq), filename);
-            add(attachment);
-            
-            // フォルダに配置する。
-            byte[] data = IOUtils.toByteArray(in);
-            Path filePath = Paths.get(putDir.toString(), filename);
-            try (BufferedOutputStream bos = new BufferedOutputStream(Files.newOutputStream(filePath))) {
-                bos.write(data, 0, data.length);
+        } else if (AttachmentPropaties.ATTACHMENT_MODE == AttachmentPropaties.AttachmentMode.DISK_MODE) {
+            // サーバのフォルダに保存する場合。
+            // ファイルを配置するフォルダを作成する。　ベースフォルダがない場合はベースフォルダごと作成する。
+            Path putDir = Paths.get(AttachmentPropaties.BASE_DIR + id + "/" + nextSeq);
+            if (Files.notExists(putDir)) {
+                try {
+                    Files.createDirectories(putDir);
+                } catch (IOException ex) {
+                    Logger.getLogger(AttachmentDb.class.getName()).log(Level.SEVERE, null, ex);
+                }
             }
-        } catch (IOException ex) {
-            Logger.getLogger(AttachmentDb.class.getName()).log(Level.SEVERE, null, ex);
+            // ファイルをフォルダに配置する。
+            try (InputStream in = part.getValueAs(InputStream.class)) {
+                // DBにデータを登録する。（添付ファイルはDBには入れずにサーバに保存する。）
+                String filename = new String(part.getFormDataContentDisposition().getFileName().getBytes("iso-8859-1"), "utf-8");
+                Attachment attachment = new Attachment(new AttachmentKey(id, nextSeq), filename);
+                add(attachment);
+
+                // フォルダに配置する。
+                byte[] data = IOUtils.toByteArray(in);
+                Path filePath = Paths.get(putDir.toString(), filename);
+                try (BufferedOutputStream bos = new BufferedOutputStream(Files.newOutputStream(filePath))) {
+                    bos.write(data, 0, data.length);
+                }
+            } catch (IOException ex) {
+                Logger.getLogger(AttachmentDb.class.getName()).log(Level.SEVERE, null, ex);
+            }
         }
     }
     
@@ -91,27 +92,29 @@ public class AttachmentDb extends TryCatchDb<Attachment> {
     public void deleteAttachment(Attachment file) {
         // DBから削除する。
         delete(file);
-        // サーバのファイルを削除する。
-        try {
-            // ファイル削除
-            Path filePath = Paths.get(AttachmentBaseDir.BASE_DIR + file.getKey().getId() + "/" + file.getKey().getSeqNo() + "/" + file.getFileName());
-            Files.deleteIfExists(filePath);
-            // 連番フォルダ削除
-            Path seqDirPath = Paths.get(AttachmentBaseDir.BASE_DIR + file.getKey().getId() + "/" + file.getKey().getSeqNo());
-            Files.deleteIfExists(seqDirPath);
-            // 番号フォルダが空の場合は削除
-            File idDir = new File(AttachmentBaseDir.BASE_DIR + file.getKey().getId());
-            if (idDir.isDirectory()) {
-                if (idDir.list().length > 0) {
-                    // フォルダが空でない場合は削除しない
-                } else {
-                    // フォルダが空の場合は削除する
-                    Path idDirPath = idDir.toPath();
-                    Files.deleteIfExists(idDirPath);
+        // DISK_MODEの場合、サーバのファイルを削除する。
+        if (AttachmentPropaties.ATTACHMENT_MODE == AttachmentPropaties.AttachmentMode.DISK_MODE) {
+            try {
+                // ファイル削除
+                Path filePath = Paths.get(AttachmentPropaties.BASE_DIR + file.getKey().getId() + "/" + file.getKey().getSeqNo() + "/" + file.getFileName());
+                Files.deleteIfExists(filePath);
+                // 連番フォルダ削除
+                Path seqDirPath = Paths.get(AttachmentPropaties.BASE_DIR + file.getKey().getId() + "/" + file.getKey().getSeqNo());
+                Files.deleteIfExists(seqDirPath);
+                // 番号フォルダが空の場合は削除
+                File idDir = new File(AttachmentPropaties.BASE_DIR + file.getKey().getId());
+                if (idDir.isDirectory()) {
+                    if (idDir.list().length > 0) {
+                        // フォルダが空でない場合は削除しない
+                    } else {
+                        // フォルダが空の場合は削除する
+                        Path idDirPath = idDir.toPath();
+                        Files.deleteIfExists(idDirPath);
+                    }
                 }
+            } catch (IOException ex) {
+                Logger.getLogger(AttachmentDb.class.getName()).log(Level.SEVERE, null, ex);
             }
-        } catch (IOException ex) {
-            Logger.getLogger(AttachmentDb.class.getName()).log(Level.SEVERE, null, ex);
         }
     }
     
@@ -141,7 +144,7 @@ public class AttachmentDb extends TryCatchDb<Attachment> {
      * @return 
      */
     public byte[] getByteFile(Attachment file) {
-        Path filePath = Paths.get(AttachmentBaseDir.BASE_DIR + file.getKey().getId() + "/" + file.getKey().getSeqNo() + "/" + file.getFileName());
+        Path filePath = Paths.get(AttachmentPropaties.BASE_DIR + file.getKey().getId() + "/" + file.getKey().getSeqNo() + "/" + file.getFileName());
         if (Files.exists(filePath)) {
             try {
                 byte[] byteFile = Files.readAllBytes(filePath);
